Extract Endpoint class into its own file
